LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStream-default-feature-policy.https-expected.txt

CONSOLE MESSAGE: line 16: The top-level frame has prevented a document with a different security origin from calling getUserMedia.
CONSOLE MESSAGE: line 16: The top-level frame has prevented a document with a different security origin from calling getUserMedia.
CONSOLE MESSAGE: line 16: The top-level frame has prevented a document with a different security origin from calling getUserMedia.
CONSOLE MESSAGE: line 16: The top-level frame has prevented a document with a different security origin from calling getUserMedia.
CONSOLE MESSAGE: line 16: The top-level frame has prevented a document with a different security origin from calling getUserMedia.
CONSOLE MESSAGE: line 16: The top-level frame has prevented a document with a different security origin from calling getUserMedia.
CONSOLE MESSAGE: line 16: The top-level frame has prevented a document with a different security origin from calling getUserMedia.


PASS Default "microphone" feature policy ["self"] allows the top-level document. 
PASS Default "microphone" feature policy ["self"] allows same-origin iframes. 
PASS Default "microphone" feature policy ["self"] disallows cross-origin iframes. 
FAIL Feature policy "microphone" can be enabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#OK" but got "#NotAllowedError"
PASS Default "camera" feature policy ["self"] allows the top-level document. 
PASS Default "camera" feature policy ["self"] allows same-origin iframes. 
PASS Default "camera" feature policy ["self"] disallows cross-origin iframes. 
FAIL Feature policy "camera" can be enabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#OK" but got "#NotAllowedError"
PASS Default "camera; microphone" feature policy ["self"] allows the top-level document. 
PASS Default "camera; microphone" feature policy ["self"] allows same-origin iframes. 
PASS Default "camera; microphone" feature policy ["self"] disallows cross-origin iframes. 
PASS Feature policy "camera; microphone" can be enabled in cross-origin iframes using "allow" attribute. 
